% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV1_10.r
\name{constraint}
\alias{constraint}
\title{Sum-to-zero constraint}
\usage{
constraint(X, S, Z = NULL)
}
\arguments{
\item{X}{A design matrix}

\item{S}{A penalty matrix or a list of penalty matrices}

\item{Z}{A list of sum-to-zero constraint matrices; default is NULL}
}
\value{
List of objects with the following items:
\item{X}{Design matrix}
\item{S}{Penalty matrix or list of penalty matrices}
\item{Z}{List of sum-to-zero constraint matrices}
}
\description{
Applies the sum-to-zero constraints to design and penalty matrices.
}
