% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV1_51.r
\name{cor.var}
\alias{cor.var}
\title{Implementation of the corrected variance Vc}
\usage{
cor.var(model)
}
\arguments{
\item{model}{survPen object, see \code{\link{survPen.fit}} for details}
}
\value{
survPen object with corrected variance Vc
}
\description{
Takes the model at convergence and calculates the variance matrix corrected for smoothing parameter uncertainty
}
