% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV1_51.r
\docType{data}
\name{datCancer}
\alias{datCancer}
\title{Patients diagnosed with cervical cancer}
\format{
A data frame with 2000 rows and 6 variables
}
\usage{
data(datCancer)
}
\description{
A simulated dataset containing the follow-up times of 2000 patients diagnosed with cervical cancer between 
1990 and 2010. End of follow-up is June 30th 2013. The variables are as follows:
\itemize{
  \item begin. beginning of follow-up. For illustration purposes about left truncation only (0--1)
  \item fu. follow-up time in years (0--5)
  \item age. age at diagnosis in years, from 21.39 to 99.33
  \item yod. decimal year of diagnosis, from 1990.023 to 2010.999
  \item dead. censoring indicator (1 for dead, 0 for censored)
  \item rate. expected mortality rate (from overall mortality of the general population) (0--0.38)
}
}
\keyword{datasets}
