% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deriv_R}
\alias{deriv_R}
\title{Derivative of a Choleski factor}
\usage{
deriv_R(deriv_Vp, p, R1)
}
\arguments{
\item{deriv_Vp}{derivatives of the Bayesian covariance matrix wrt rho (log smoothing parameters).}

\item{p}{number of regression parameters}

\item{R1}{Choleski factor of Vp}
}
\value{
a list containing the derivatives of R1 wrt rho (log smoothing parameters)
}
\description{
Derivative of a Choleski factor
}
