% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV2_00.r
\docType{data}
\name{expected.table}
\alias{expected.table}
\title{French women mortality table}
\format{
A data frame with 22866 rows and 3 variables
}
\usage{
data(expected.table)
}
\description{
French women mortality table to serve as example of reference/expected mortality in excess hazard and relative mortality ratio models
The data come from the human mortality databse website: https://www.mortality.org/Country/Country?cntr=FRATNP
\itemize{
  \item Age. Age group for 1-year interval from exact age x to just before exact age x+1 (0-110+)
  \item Year. Calendar Year (1816-2021)
  \item mx. Central death rate between ages x and x+1
}
}
\keyword{datasets}
