% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{\%cross\%}
\alias{\%cross\%}
\title{Matrix cross-multiplication between two matrices}
\usage{
Mat1 \%cross\% Mat2
}
\arguments{
\item{Mat1}{a matrix.}

\item{Mat2}{another matrix.}
}
\value{
prod the product t(Mat1)%*%Mat2
}
\description{
Matrix cross-multiplication between two matrices
}
