% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{\%mult\%}
\alias{\%mult\%}
\title{Matrix multiplication between two matrices}
\usage{
Mat1 \%mult\% Mat2
}
\arguments{
\item{Mat1}{a matrix.}

\item{Mat2}{another matrix.}
}
\value{
prod the product Mat1%*%Mat2
}
\description{
Matrix multiplication between two matrices
}
