% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{\%vec\%}
\alias{\%vec\%}
\title{Matrix multiplication between a matrix and a vector}
\usage{
Mat \%vec\% vec
}
\arguments{
\item{Mat}{a matrix.}

\item{vec}{a vector.}
}
\value{
prod the product Mat%*%vec
}
\description{
Matrix multiplication between a matrix and a vector
}
