% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HazGL}
\alias{HazGL}
\title{Gauss-Legendre evaluations}
\usage{
HazGL(X_GL, n_legendre, beta)
}
\arguments{
\item{X_GL}{list of matrices (\code{length(X.GL)=n.legendre}) for Gauss-Legendre quadrature}

\item{n_legendre}{number of nodes for Gauss-Legendre quadrature}

\item{beta}{vector of estimated regression parameters}
}
\value{
list of all the matrix-vector multiplications X.GL[[i]]\%*\%beta for Gauss Legendre integration in order to save computation time
}
\description{
Gauss-Legendre evaluations
}
