% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV2_00.r
\docType{data}
\name{HeartFailure}
\alias{HeartFailure}
\title{Patients with heart failure at risk of recurrent hospitalization events}
\format{
A data frame with 3 068 rows and 6 variables
}
\usage{
data(HeartFailure)
}
\description{
A simulated dataset containing 3 068 observations (2 268 events) in 800 patients with heart failure. 
The dataset is based on hfaction_cpx12 dataset from package WA. The variables are as follows:
\itemize{
  \item id. patient identifcation number
  \item treatment. treatment=0 for control and treatment=1 for exercise training
  \item t0. beginning of follow-up for a given event
  \item t1. end of follow-up for a given event (up to 3.27 years)
  \item enum. event identification number for a given patient (between 1 and 6 events per patient)
  \item event. event indicator (1 for hospitalization, 0 for censored)
}
}
\keyword{datasets}
