% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV1_51.r
\name{inv.repam}
\alias{inv.repam}
\title{Reverses the initial reparameterization for stable evaluation of the log determinant of the penalty matrix}
\usage{
inv.repam(model, X.ini, S.pen.ini)
}
\arguments{
\item{model}{survPen object, see \code{\link{survPen.fit}} for details}

\item{X.ini}{initial design matrix (before reparameterization)}

\item{S.pen.ini}{initial penalty matrices}
}
\value{
survPen object with standard parameterization
}
\description{
Transforms the final model by reversing the initial reparameterization performed by \code{\link{repam}}. Derives the corrected version of the Bayesian covariance matrix
}
