% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV1_50.r
\name{print.summary.survPen}
\alias{print.summary.survPen}
\title{print summary for a \code{survPen} fit}
\usage{
\method{print}{summary.survPen}(x, ...)
}
\arguments{
\item{x}{an object of class \code{summary.survPen}}

\item{...}{other arguments}
}
\value{
print of summary
}
\description{
print summary for a \code{survPen} fit
}
