% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV1_51.r
\name{rd}
\alias{rd}
\title{Defining random effects in survPen formulae}
\usage{
rd(...)
}
\arguments{
\item{...}{Any number of covariates separated by ","}
}
\value{
object of class \code{rd.smooth.spec}
}
\description{
Used inside a formula object to define a random effect.
}
\examples{
# cubic regression spline of time with 10 unspecified knots + random effect at the cluster level
formula.test <- ~smf(time,df=10) + rd(cluster)


}
