% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV1_10.r
\name{repam}
\alias{repam}
\title{Applies initial reparameterization for stable evaluation of the log determinant of the penalty matrix}
\usage{
repam(build)
}
\arguments{
\item{build}{object as returned by \code{\link{model.cons}}}
}
\value{
\item{build}{an object as returned by \code{\link{model.cons}}}
\item{X.ini}{initial design matrix (before reparameterization)}
\item{S.pen.ini}{initial penalty matrices}
}
\description{
Transforms the object from \code{\link{model.cons}} by applying the matrix reparameterization (matrix U.F). The reparameterization
is reversed at convergence by \code{\link{inv.repam}}.
}
