% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV2_00.r
\name{robust.var}
\alias{robust.var}
\title{Implementation of the robust variance Vr}
\usage{
robust.var(model, data, cluster.name, n.legendre = 50)
}
\arguments{
\item{model}{survPen object, see \code{\link{survPen.fit}} for details}

\item{data}{original dataset}

\item{cluster.name}{name of cluster variable in data}

\item{n.legendre}{number of nodes for Gauss-Legendre quadrature; default is 50}
}
\value{
survPen object with robust variance Vr
}
\description{
Takes the model at convergence and calculates the robust variance matrix accounting for correlated survival times
}
