% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV1_10.r
\name{tensor.prod.S}
\alias{tensor.prod.S}
\title{Tensor product for penalty matrices}
\usage{
tensor.prod.S(S)
}
\arguments{
\item{S}{list of m marginal penalty matrices}
}
\value{
\item{TS}{List of the penalty matrices associated with the tensor product smooth}
}
\description{
Computes the penalty matrices of a tensor product smooth from the marginal penalty matrices. The code is from
function \code{tensor.prod.penalties} in \code{mgcv} package.
}
