% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV1_10.r
\name{tensor.prod.X}
\alias{tensor.prod.X}
\title{tensor model matrix}
\usage{
tensor.prod.X(X)
}
\arguments{
\item{X}{list of m design matrices with n rows and p1, p2, ... pm columns respectively}
}
\value{
\item{T}{Matrix of dimensions n*(p1*p2*...*pm) representing the row tensor product of the matrices in X}
}
\description{
Computes the model matrix of tensor product smooth from the marginal bases.
}
