% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPenV1_51.r
\name{print.summary.survPen}
\alias{print.summary.survPen}
\title{print summary for a \code{survPen} fit}
\usage{
\method{print}{summary.survPen}(
  x,
  digits = max(3, getOption("digits") - 2),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{an object of class \code{summary.survPen}}

\item{digits}{controls number of digits printed in output.}

\item{signif.stars}{Should significance stars be printed alongside output.}

\item{...}{other arguments}
}
\value{
print of summary
}
\description{
print summary for a \code{survPen} fit
}
